pyproject.toml
[project]
name = "winter-teamf-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "woohyun23", email = "parkwh0823@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.1.4,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "mysqlclient (>=2.2.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pymupdf4llm (>=0.0.17,<0.0.18)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "openai (>=0.28.0,<1.0.0)",
    "fpdf (>=1.7.2,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pinecone-client (>=5.0.1,<6.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "reportlab (>=4.2.5,<5.0.0)",
    "celery[rabbitmq] (>=5.4.0,<6.0.0)",
    "langchain-community (>=0.3.14,<0.4.0)"

]

packages = ["winter_teamf_backend"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
